requirements.txt
alembic==1.18.3
bcrypt==4.1.3
blinker==1.9.0
cachelib==0.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.20.0
packaging==26.0
pillow==12.1.0
pycparser==3.0
PyMySQL==1.1.1
python-dotenv==1.0.1
qrcode==8.2
redis==5.0.7
reportlab==4.4.9
requests==2.32.5
setuptools==80.10.2
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.0.3
wheel==0.46.3
win32_setctime==1.2.0
WooCommerce==3.0.0
